chore: move JobFilterBat to /jobs
