+++ src/app/(student)/jobs/JobFilterBar.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import { Input } from "./ui/input";
+import { Input } from "../../../components/ui/input";
 import {
     Select,
     SelectContent,
